secure-api/requirements.txt
fastapi
uvicorn
groq
openai
python-multipart
cryptography
mariadb
pyjwt
passlib[bcrypt]
pycryptodome
mysql-connector-python
